Major refactor, add completed specialization highlight
Moved lots of course handling stuff from the controllers to the new
serialization class as it was getting harder to manage communication
between controllers. Added top-level angular update service for the
checkboxes. Moved parsing into cookies as it is only used for cookies.
Removed _completed altogether, only stored in cookies from now on.